Modified returntype for functions getNodeCount, getLevel, getMaxTreeDepth, getMaxLevel, getTreeDepth from int -> UINT (classes ExpressionNode,ParserTree, NodeCounter, MaxLevelFinder) Added function UINT getHeight() const to SNode Modified ParserTree::treeToString() to print TreeForm, State and reduceIteration

+++ Lib/Src/Expression/ParserTreeComplexity.cpp
@@ -10,9 +10,9 @@ ParserTreeComplexity::ParserTreeComplexity(const ParserTree &tree) {
 }
 
 int ParserTreeComplexity::compare(const ParserTreeComplexity &tc) const {
-  const int c = (m_nodeCount + 2 * m_nameCount) - (tc.m_nodeCount + 2 * tc.m_nameCount);
+  const int c = (int)(m_nodeCount + 2 * m_nameCount) - (int)(tc.m_nodeCount + 2 * tc.m_nameCount);
   if(c) return c;
-  return m_treeDepth - tc.m_treeDepth;
+  return (int)m_treeDepth - (int)tc.m_treeDepth;
 }
 
 ParserTreeComplexity ParserTree::getComplexity() const {
